Creating modules to handle image and reddit
